refactor(auth api): rename class name
