+++ apollo-portal/src/test/java/com/ctrip/framework/apollo/portal/service/ConfigsExportServiceTest.java
@@ -23,7 +23,7 @@ import com.ctrip.framework.apollo.common.entity.App;
 import com.ctrip.framework.apollo.common.entity.AppNamespace;
 import com.ctrip.framework.apollo.core.enums.ConfigFileFormat;
 import com.ctrip.framework.apollo.portal.AbstractUnitTest;
-import com.ctrip.framework.apollo.portal.component.PermissionValidator;
+import com.ctrip.framework.apollo.portal.component.UserPermissionValidator;
 import com.ctrip.framework.apollo.portal.entity.bo.ItemBO;
 import com.ctrip.framework.apollo.portal.entity.bo.NamespaceBO;
 import com.ctrip.framework.apollo.portal.entity.bo.UserInfo;
@@ -65,7 +65,7 @@ public class ConfigsExportServiceTest extends AbstractUnitTest {
   @Mock
   private NamespaceService          namespaceService;
   @Mock
-  private PermissionValidator       permissionValidator;
+  private UserPermissionValidator userPermissionValidator;
   @Mock
   private UserInfoHolder            userInfoHolder;
   @Mock
@@ -155,7 +155,7 @@ public class ConfigsExportServiceTest extends AbstractUnitTest {
 
     when(appService.findAll()).thenReturn(exportApps);
     when(appNamespaceService.findAll()).thenReturn(appNamespaces);
-    when(permissionValidator.isAppAdmin(any())).thenReturn(true);
+    when(userPermissionValidator.isAppAdmin(any())).thenReturn(true);
     when(clusterService.findClusters(env, appId1)).thenReturn(app1Clusters);
     when(clusterService.findClusters(env, appId2)).thenReturn(app2Clusters);
     when(namespaceService.findNamespaceBOs(appId1, Env.DEV, clusterName1, fillItemDetail, false)).thenReturn(app1Cluster1Namespace);
